Refactor Blog entity and update URL property config
Modified ApplicationDbContext.cs to set the URL property in the Blog entity to varchar(150), marked it as required, and set a max length of 200 characters. Added a comment to the URL column in the database. In Blog.cs, removed the using directive for System.ComponentModel.DataAnnotations, adjusted the namespace declaration, and commented out attributes for the URL and Posts properties.